Heater control

+++ src/hardware/SensorDHT22.h
@@ -7,12 +7,12 @@
 class SensorDHT22
 {
     public:
-        SensorDHT22(int pin);
+        SensorDHT22(uint8_t pin);
         void init();
-        float getTemp();
-        float getHumi();
+        int8_t getTemp();
+        int8_t getHumi();
     private:
-        int _pin;
+        uint8_t _pin;
         DHT* _dht;
 };
